backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-resized==1.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2025.3
